fix: Synchronize API requests
Turned out that the Hetzner DNS API doesn't like multiple DNS records
created at the same time. And if there is now implicit dependency
between two resources, Terraform assumes it is safe to create more than
one resource at a time.

There is a bulk create endpoint, which might be the "right" way to
implement the resource. However, it would be neccessary to add a DNS
record group resource or something that groups a list of DNS records and
this make the CRUD operations of this resource more complicated.

I think having one resource per DNS record is good for now. Added a
Mutex to synchronize all API requests. I hope this is an acceptable
slow down. If there is demand for a DNS record set resources, let me
know. It shouldn't be hard to and probably a good first issue for a
contributer :-).

fixes #5