Document how to import records
Decided for importing using record id. Terraform import allows only one parameter, which is the record id. In order to import a record by name, the zone is required as well. Both could be encoded in one string and used as a parameter. However, looking up the id using the Hetzner API manually is good enough for now.